perf(runtime): shrink DuxStr header to 16 bytes via int32_t len
Replacing int64_t len with int32_t len eliminates the 4-byte alignment
padding that followed refcount, shrinking sizeof(DuxStr) from 24 to 16:

  old: { i32 refcount, [4 pad], i64 len, ptr ext, char[] }  sizeof=24
  new: { i32 refcount,  i32 len,          ptr ext, char[] }  sizeof=16

Short strings (len <= DUXSTR_INLINE_MAX) now allocate malloc(16+len+1),
landing in the same glibc size-class as a pure-FAM struct (which also
has a 16-byte header).  Previously the 24-byte header pushed short strings
into a larger bin, explaining the speed gap.

Benchmark results vs two-alloc baseline (2 M ops, -O2, fair ->len reads):

  len=8:   hybrid (inline) ≈ fam-only   (+63–79%  vs two-alloc)
  len=40:  hybrid (inline)  ~ fam-only   (+74–77%  vs two-alloc)
  len=64:  hybrid (ext)      -3 to -7%   (two-alloc path, expected)
  len=200: hybrid (ext)      -3 to +2%   (beats fam-only by ~20%)

The remaining small gap at len=40 is the irreducible cost of the hybrid:
one extra store (ext=NULL) and one conditional branch in release.
Max string length is now 2 GiB (int32_t), which is sufficient for Dux.

Changes:
- duxrt.h: int32_t len field, updated sizeof comment
- str.c: cast to (int32_t) on store; explicit (int64_t) on public returns
- codegen.cpp: str_literal() emits { i32, i32, ptr, [N+1 x i8] } so the
               LLVM global layout stays aligned with the C struct (data
               at offset 16 instead of 24)
- benchmarks/str_alloc_bench.c: HybStr updated to int32_t len; all three
               variants now read ->len for a fair comparison

+++ src/runtime/str.c
@@ -27,7 +27,7 @@ DuxStr* duxrt_str_new(const char* src, int64_t len) {
         s->ext[len] = '\0';
     }
     s->refcount = 1;
-    s->len      = len;
+    s->len      = (int32_t)len;   /* int32_t field; caller ensures len <= INT32_MAX */
     return s;
 }
 
@@ -74,7 +74,7 @@ DuxStr* duxrt_str_from_double(double v) {
 }
 
 int64_t duxrt_str_length(DuxStr* s) {
-    return s ? s->len : 0;
+    return s ? (int64_t)s->len : 0;
 }
 
 DuxStr* duxrt_str_index(DuxStr* s, int64_t i) {
@@ -100,5 +100,5 @@ int duxrt_str_eq(DuxStr* a, DuxStr* b) {
 }
 
 int64_t duxrt_len(DuxStr* s) {
-    return s ? s->len : 0;
+    return s ? (int64_t)s->len : 0;
 }
